add validation path to polywrap.test.yaml
