+++ packages/cli/lang/es.json
@@ -154,9 +154,11 @@
   "commands_run_options_validateScript": "Validate the output of the workflow jobs",
   "commands_run_options_validate": "cue-file",
   "commands_run_options_outputFile": "Output file path for the workflow result",
+  "commands_run_options_manifest": "manifest",
   "commands_run_options_quiet": "Suppress output",
   "commands_run_options_jobIds": "jobs",
   "commands_run_options_jobs": "Specify ids of jobs that you want to run",
+  "commands_run_manifestPathDescription": "Workflow Manifest path",
   "commands_run_error_validatorNotFound": "validate script not found at: {path}",
   "commands_run_error_outputFileMissing": "{option} option missing {argument} argument",
   "commands_run_error_clientConfigMissingPath": "{option} option missing {argument} argument",
