+++ packages/cli/src/commands/run.ts
@@ -2,16 +2,15 @@ import { Command, Program } from "./types";
 import {
   intlMsg,
   parseClientConfigOption,
-  parseValidateScriptOption,
   parseWorkflowOutputFilePathOption,
   validateOutput,
   JobRunner,
   JobStatus,
   JobResult,
+  loadWorkflowManifest,
 } from "../lib";
 
 import { Uri } from "@polywrap/core-js";
-import { PolywrapWorkflow } from "@polywrap/polywrap-manifest-types-js";
 import { PolywrapClient, PolywrapClientConfig } from "@polywrap/client-js";
 import path from "path";
 import yaml from "js-yaml";
@@ -19,8 +18,9 @@ import fs from "fs";
 
 type WorkflowCommandOptions = {
   clientConfig: Partial<PolywrapClientConfig>;
+  manifest: string;
   jobs?: string[];
-  validateScript?: string;
+  validationScript?: string;
   outputFile?: string;
   quiet?: boolean;
 };
@@ -31,17 +31,14 @@ export const run: Command = {
       .command("run")
       .alias("r")
       .description(intlMsg.commands_run_description())
-      .argument(
-        `<${intlMsg.commands_run_options_workflow()}>`,
-        intlMsg.commands_run_options_workflowScript()
+      .option(
+        `-m, --manifest  <${intlMsg.commands_run_options_manifest()}>`,
+        intlMsg.commands_run_manifestPathDescription(),
+        "polywrap.test.yaml"
       )
       .option(
         `-c, --client-config <${intlMsg.commands_common_options_configPath()}>`,
         `${intlMsg.commands_common_options_config()}`
-      )
-      .option(
-        `-v, --validate-script <${intlMsg.commands_run_options_validate()}>`,
-        `${intlMsg.commands_run_options_validateScript()}`
       )
       .option(
         `-o, --output-file <${intlMsg.commands_run_options_outputFilePath()}>`,
@@ -52,34 +49,43 @@ export const run: Command = {
         intlMsg.commands_run_options_jobs()
       )
       .option(`-q, --quiet`, `${intlMsg.commands_run_options_quiet()}`)
-      .action(async (workflow: string, options) => {
-        await _run(workflow, {
+      .action(async (options) => {
+        await _run({
           ...options,
           clientConfig: await parseClientConfigOption(options.clientConfig),
           outputFile: options.outputFile
             ? parseWorkflowOutputFilePathOption(options.outputFile)
-            : undefined,
-          validateScript: options.validateScript
-            ? parseValidateScriptOption(options.validateScript)
             : undefined,
         });
       });
   },
 };
 
-const _run = async (workflowPath: string, options: WorkflowCommandOptions) => {
-  const { clientConfig, outputFile, validateScript, quiet, jobs } = options;
-  const client = new PolywrapClient(clientConfig);
+function loadValidationScript(cueFilepath: string): string {
+  cueFilepath = path.resolve(cueFilepath);
 
-  function getParser(path: string) {
-    return path.endsWith(".yaml") || path.endsWith(".yml")
-      ? yaml.load
-      : JSON.parse;
+  if (!fs.existsSync(cueFilepath)) {
+    console.error(
+      intlMsg.commands_run_error_validatorNotFound({
+        path: cueFilepath,
+      })
+    );
+    process.exit(1);
   }
 
-  const workflow: PolywrapWorkflow = getParser(workflowPath)(
-    fs.readFileSync(workflowPath).toString()
-  );
+  return cueFilepath;
+}
+
+const _run = async (options: WorkflowCommandOptions) => {
+  const { manifest, clientConfig, outputFile, quiet, jobs } = options;
+  const client = new PolywrapClient(clientConfig);
+
+  const manifestPath = path.resolve(manifest);
+  const workflow = loadWorkflowManifest(manifestPath);
+  const validationScript = workflow.validation
+    ? loadValidationScript(workflow.validation)
+    : undefined;
+
   const workflowOutput: (JobResult & { id: string })[] = [];
 
   const onExecution = async (id: string, jobResult: JobResult) => {
@@ -100,8 +106,8 @@ const _run = async (workflowPath: string, options: WorkflowCommandOptions) => {
       process.exitCode = 1;
     }
 
-    if (status !== JobStatus.SKIPPED && validateScript) {
-      await validateOutput(id, { data, error }, validateScript, quiet);
+    if (status !== JobStatus.SKIPPED && validationScript) {
+      await validateOutput(id, { data, error }, validationScript, quiet);
     }
 
     if (!quiet) {
