+++ packages/cli/src/lib/option-parsers/run.ts
@@ -16,18 +16,3 @@ export function parseWorkflowScriptPathOption(script: string): string {
 export function parseWorkflowOutputFilePathOption(outputFile: string): string {
   return path.resolve(outputFile);
 }
-
-export function parseValidateScriptOption(cueFile: string): string {
-  const cueFilepath = path.resolve(cueFile);
-
-  if (!fs.existsSync(cueFilepath)) {
-    console.error(
-      intlMsg.commands_run_error_validatorNotFound({
-        path: cueFilepath,
-      })
-    );
-    process.exit(1);
-  }
-
-  return cueFilepath;
-}
