+++ packages/manifests/polywrap/formats/polywrap.test/0.1.0.json
@@ -14,6 +14,10 @@
       "type": "string",
       "pattern": "^[a-zA-Z0-9\\-\\_]+$"
     },
+    "validation": {
+      "description": "Path to validation script",
+      "type": "string"
+    },
     "jobs": {
       "$ref": "#/definitions/jobs"
     }
