+++ packages/templates/wasm/rust/package.json
@@ -11,7 +11,7 @@
     "test": "yarn test:e2e && yarn test:workflow",
     "test:e2e": "yarn test:e2e:codegen && jest --passWithNoTests --runInBand --verbose",
     "test:e2e:codegen": "npx polywrap app codegen -m ./src/__tests__/types/polywrap.app.yaml -g ./src/__tests__/types/wrap",
-    "test:workflow": "npx polywrap run -o ./workflows/output.json -v ./workflows/validator.cue ./workflows/e2e.yaml"
+    "test:workflow": "npx polywrap run -o ./workflows/output.json"
   },
   "devDependencies": {
     "@types/jest": "27.0.3",
