+++ packages/test-cases/cases/cli/wasm/run/polywrap.test.invalid.json
@@ -1,5 +1,7 @@
 {
   "name-invalid": "simple-storage",
+  "format": "0.1",
+  "validation": "./validator.cue",
   "jobs": {
     "cases": {
       "steps": [
