+++ packages/test-cases/cases/cli/wasm/run/validator.cue
